Change 2bme type to double

+++ lib/TwoBodyOperator.h
@@ -17,9 +17,9 @@ class TwoBodyOperatorChannel : public TwoBodySpace {
         TwoBodyOperatorChannel(){};
         TwoBodyOperatorChannel(TwoBodyChannel chbra1, TwoBodyChannel chket1);
 
-        int get_2bme(int idxbra, int idxket);
-        int get_2bme_orbit_indices(int a, int b, int c, int d);
-        int get_2bme_orbits(Orbit oa, Orbit ob, Orbit oc, Orbit od);
+        double get_2bme(int idxbra, int idxket);
+        double get_2bme_orbit_indices(int a, int b, int c, int d);
+        double get_2bme_orbits(Orbit oa, Orbit ob, Orbit oc, Orbit od);
         void set_2bme(int idxbra, int idxket, double v);
         void set_2bme_orbit_indices(int a, int b, int c, int d, double v);
         void set_2bme_orbits(Orbit oa, Orbit ob, Orbit oc, Orbit od, double v);
@@ -37,11 +37,11 @@ class TwoBodyOperator : public TwoBodyOperatorChannel {
         TwoBodyOperator(){};
         TwoBodyOperator(TwoBodySpace two_body_space1, int rankJ1=0, int rankP1=1, int rankZ1=0);
 
-        int get_2bme(int ichbra, int ichket, int idxbra, int idxket);
-        int get_2bme_orbit_indices(int ichbra, int ichket, int a, int b, int c, int d);
-        int get_2bme_orbits(int ichbra, int ichket, Orbit oa, Orbit ob, Orbit oc, Orbit od);
-        int get_2bme_orbitsJ(Orbit oa, Orbit ob, Orbit oc, Orbit od, int Jab, int Jcd);
-        int get_2bme_orbit_indicesJ(int a, int b, int c, int d, int Jab, int Jcd);
+        double get_2bme(int ichbra, int ichket, int idxbra, int idxket);
+        double get_2bme_orbit_indices(int ichbra, int ichket, int a, int b, int c, int d);
+        double get_2bme_orbits(int ichbra, int ichket, Orbit oa, Orbit ob, Orbit oc, Orbit od);
+        double get_2bme_orbitsJ(Orbit oa, Orbit ob, Orbit oc, Orbit od, int Jab, int Jcd);
+        double get_2bme_orbit_indicesJ(int a, int b, int c, int d, int Jab, int Jcd);
 
         void set_2bme(int ichbra, int ichket, int idxbra, int idxket, double v);
         void set_2bme_orbit_indices(int ichbra, int ichket, int a, int b, int c, int d, double v);
